+++ packages/browser/src/extensions/sentry-integration.ts
@@ -15,6 +15,7 @@
  * @param {Number} [projectId] Optional: The Sentry project id, used to send a direct link from PostHog to Sentry
  * @param {string} [prefix] Optional: Url of a self-hosted sentry instance (default: https://sentry.io/organizations/)
  * @param {SeverityLevel[] | '*'} [severityAllowList] Optional: send events matching the provided levels. Use '*' to send all events (default: ['error'])
+ * @param {boolean} [sendExceptionsToPostHog] Optional: capture exceptions as events in PostHog (default: true)
  */
 
 import { PostHog } from '../posthog-core'
@@ -63,13 +64,20 @@ export type SentryIntegrationOptions = {
     projectId?: number
     prefix?: string
     severityAllowList?: SeverityLevel[] | '*'
+    sendExceptionsToPostHog?: boolean
 }
 
 const NAME = 'posthog-js'
 
 export function createEventProcessor(
     _posthog: PostHog,
-    { organization, projectId, prefix, severityAllowList = ['error'] }: SentryIntegrationOptions = {}
+    {
+        organization,
+        projectId,
+        prefix,
+        severityAllowList = ['error'],
+        sendExceptionsToPostHog = true,
+    }: SentryIntegrationOptions = {}
 ): (event: _SentryEvent) => _SentryEvent {
     return (event) => {
         const shouldProcessLevel = severityAllowList === '*' || severityAllowList.includes(event.level as SeverityLevel)
@@ -135,7 +143,9 @@ export function createEventProcessor(
                 event.event_id
         }
 
-        _posthog.exceptions.sendExceptionEvent(data)
+        if (sendExceptionsToPostHog) {
+            _posthog.exceptions.sendExceptionEvent(data)
+        }
 
         return event
     }
@@ -165,13 +175,20 @@ export class SentryIntegration implements _SentryIntegrationClass {
         organization?: string,
         projectId?: number,
         prefix?: string,
-        severityAllowList?: SeverityLevel[] | '*'
+        severityAllowList?: SeverityLevel[] | '*',
+        sendExceptionsToPostHog?: boolean
     ) {
         // setupOnce gets called by Sentry when it intializes the plugin
         this.name = NAME
         this.setupOnce = function (addGlobalEventProcessor: (callback: _SentryEventProcessor) => void) {
             addGlobalEventProcessor(
-                createEventProcessor(_posthog, { organization, projectId, prefix, severityAllowList })
+                createEventProcessor(_posthog, {
+                    organization,
+                    projectId,
+                    prefix,
+                    severityAllowList,
+                    sendExceptionsToPostHog: sendExceptionsToPostHog ?? true,
+                })
             )
         }
     }