+++ packages/core/jest.config.mjs
@@ -1,0 +1,11 @@
+export default {
+  moduleNameMapper: {
+    '^@/(.*)$': '<rootDir>/src/$1',
+  },
+  collectCoverage: true,
+  clearMocks: true,
+  fakeTimers: { enableGlobally: true },
+  coverageDirectory: 'coverage',
+  silent: true,
+  verbose: false,
+}