+++ packages/core/src/index.ts
@@ -35,7 +35,7 @@ import {
   allSettled,
   assert,
   currentISOTime,
-  isError,
+  isPlainError,
   removeTrailingSlash,
   retriable,
   RetriableOptions,
@@ -50,6 +50,7 @@ import { PromiseQueue } from './utils/promise-queue'
 
 export { getFeatureFlagValue } from './featureFlagUtils'
 export * from './utils'
+export * as ErrorTracking from './error-tracking'
 
 class PostHogFetchHttpError extends Error {
   name = 'PostHogFetchHttpError'
@@ -2116,8 +2117,8 @@ export abstract class PostHogCore extends PostHogCoreStateless {
       $exception_level: 'error',
       $exception_list: [
         {
-          type: isError(error) ? error.name : 'Error',
-          value: isError(error) ? error.message : error,
+          type: isPlainError(error) ? error.name : 'Error',
+          value: isPlainError(error) ? error.message : error,
           mechanism: {
             handled: true,
             synthetic: false,